Abstract parser logic
